backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
APScheduler>=3.10.0
requests>=2.28.0
resend>=2.0.0
twilio>=8.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
openai>=1.0.0
